requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
bcrypt==4.1.2
email-validator==2.1.0
flasgger==0.9.7.1
marshmallow==3.20.1
requests==2.31.0
werkzeug==3.0.1
